+++ src/helpers/generateEnums.js
@@ -3,7 +3,7 @@ module.exports = function generateEnums(schemas) {
   return Object.keys(schemas)
     .flatMap((schemaName) => {
       const schema = schemas[schemaName];
-      return Object.keys(schema.properties)
+      return schema.properties ? Object.keys(schema.properties)
         .filter((property) => schema.properties[property].enum)
         .map((property) => {
           const enumValues = schema.properties[property].enum;
@@ -11,7 +11,7 @@ module.exports = function generateEnums(schemas) {
           const enumDefinition = `
 export type ${enumName} = ${enumValues.map((item) => `'${item}'`).join(" | ")};`;
           return enumDefinition.trim();
-        });
+        }) : [''];
     })
     .join("\n\n");
 };